feat: Improve startup session seeding and multi-target notifications. Include GPU names in startup summaries and service formatters. Allow notification formatters to return multi-message payloads and chunk Discord summaries safely under 2k characters. Capture current rental counters in snapshots and split placeholders across active sessions when seeding rentals.

+++ Classes/vast_machine.py
@@ -61,7 +61,7 @@ class VastMachine(BaseModel):
     direct_port_count: int
     public_ipaddr: str
     num_reports: Optional[int]
-    num_recent_reports: Optional[int]
+    num_recent_reports: Optional[float]
     client_end_date: Optional[float]
     machine_maintenance: Optional[str]
     driver_version: str
